backend/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0

# LangGraph & LangChain
langgraph>=0.0.40
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20

# LLM API Clients
groq>=0.4.0
anthropic>=0.18.0
httpx>=0.26.0

# Computer Vision
ultralytics>=8.1.0
opencv-python>=4.9.0.80
Pillow>=10.2.0

# OCR
pytesseract>=0.3.10
easyocr>=1.7.1

# PDF Processing
pdf2image>=1.17.0
pdfplumber>=0.10.4
PyPDF2>=3.0.1

# Vector Database
qdrant-client>=1.7.0

# Embeddings
sentence-transformers>=2.3.1
transformers>=4.37.0
torch>=2.1.0

# Data Processing
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.12.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiofiles>=23.2.1
python-magic>=0.4.27

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0

# Development
black>=24.1.0
isort>=5.13.0
mypy>=1.8.0
